Renaming interface_status
