--- test_modules/interface_status/go.mod
@@ -1,7 +1,0 @@
-module interface_status
-
-go 1.15
-
-require github.com/buger/jsonparser v1.0.0
-require github.com/ccoverstreet/Jablko v0.1.0
-replace github.com/ccoverstreet/Jablko => ../../../Jablko
