+++ test_modules/interfacestatus/go.mod
@@ -1,0 +1,10 @@
+module interface_status
+
+go 1.15
+
+require (
+	github.com/buger/jsonparser v1.0.0
+	github.com/ccoverstreet/Jablko v0.0.0-00010101000000-000000000000 // indirect
+)
+
+replace github.com/ccoverstreet/Jablko => ../../../Jablko
